An exploding endercrystal removes other ecs
the other ecs won't do an explosion though
Had to move the damage() methods up into GlowEntity
because multiple non living entities can also get damages
like minecarts, boats, endercrystals
but they do not have, in the most cases, have health in the same sense
as a living entity hase
