+++ src/main/java/net/glowstone/entity/GlowLivingEntity.java
@@ -704,21 +704,6 @@ public abstract class GlowLivingEntity extends GlowEntity implements LivingEntit
     }
 
     @Override
-    public void damage(double amount) {
-        damage(amount, null, DamageCause.CUSTOM);
-    }
-
-    @Override
-    public void damage(double amount, Entity source) {
-        damage(amount, source, DamageCause.CUSTOM);
-    }
-
-    @Override
-    public void damage(double amount, DamageCause cause) {
-        damage(amount, null, cause);
-    }
-
-    @Override
     public void damage(double amount, Entity source, DamageCause cause) {
         // invincibility timer
         if (noDamageTicks > 0 || health <= 0 || !canTakeDamage(cause) || isInvulnerable()) {
